+++ src/components/Posts/partials/SingleComment/styles.jsx
