+++ src/pages/Home/partials/Main/index.jsx
@@ -1,15 +1,15 @@
-import { useEffect, useState } from 'react';
-import { connect } from 'react-redux';
-import { getArticlesAPI } from '../../../../services/actions';
+import { useEffect, useState } from "react";
+import { connect } from "react-redux";
+import { getArticlesAPI } from "../../../../services/actions";
 
-import { IoMdSettings } from 'react-icons/io';
-import { Container, Content, PostNav } from './styles';
+import { IoMdSettings } from "react-icons/io";
+import { Container, Content, PostNav } from "./styles";
 
-import PostModal from '../../../../components/Posts/PostModal';
-import { SinglePost } from '../../../../components/Posts/SinglePost';
+import PostModal from "../../../../components/Posts/PostModal";
+import { SinglePost } from "../../../../components/Posts/SinglePost";
 
 const Main = (props) => {
-  const [showModal, setShowModal] = useState('close');
+  const [showModal, setShowModal] = useState("close");
   useEffect(() => {
     props.getArticles();
   }, []);
@@ -21,46 +21,17 @@ const Main = (props) => {
     }
 
     switch (showModal) {
-      case 'open':
-        setShowModal('close');
+      case "open":
+        setShowModal("close");
         break;
-      case 'close':
-        setShowModal('open');
+      case "close":
+        setShowModal("open");
         break;
       default:
-        setShowModal('close');
+        setShowModal("close");
         break;
     }
   };
-
-  async function generateUrl(url, callback) {
-    const xhr = new XMLHttpRequest();
-    xhr.onload = () => {
-      var reader = new FileReader();
-      reader.onloadend = () => {
-        callback(reader.result);
-      };
-      reader.readAsDataURL(xhr.response);
-    };
-    xhr.open('GET', url);
-    xhr.responseType = 'blob';
-    xhr.send();
-  }
-
-  function setArticleImages(article) {
-    const urlList = [];
-    // article.forEach((article) => {
-    //   if (article.sharedImg) {
-    //     (async () => {
-    //       const file = await generateUrl(article.sharedImg, () => {
-
-    //       });
-    //       console.log(file);
-    //     })();
-    //   }
-    // });
-    return true;
-  }
 
   return (
     <Container>
@@ -68,24 +39,23 @@ const Main = (props) => {
         <button
           onClick={handleClick}
           disabled={!!props.loading}
-          className='post-space'
+          className="post-space"
         >
           Start a Post
         </button>
         <IoMdSettings
           size={30}
-          style={{ margin: '8px', fill: 'whitesmoke' }}
+          style={{ margin: "8px", fill: "whitesmoke" }}
         ></IoMdSettings>
       </PostNav>
       <Content>
-        {props.loading && <img src='public/images/spin-loading.gif' alt='' />}
+        {props.loading && <img src="public/images/spin-loading.gif" alt="" />}
         {props.articles.length != 0 &&
-          setArticleImages(props.articles) &&
           props.articles.map((article) => (
             <SinglePost article={article} id={article.id} key={article.id} />
           ))}
       </Content>
-      {showModal === 'open' && (
+      {showModal === "open" && (
         <PostModal showModal={showModal} handleClick={handleClick} />
       )}
     </Container>
