+++ src/styles/utilities.css
@@ -1,4 +1,15 @@
+/* for home page */
 .welcome {
   text-decoration: none;
   color: #00000090;
+}
+
+/* for use in posts */
+.menu_button {
+  fill: #99d3df;
+  transform: rotate(90deg) scale(0.5);
+}
+
+.menu_items {
+  width: 80%;
 }
